Fix crashing issue in benchmark runner caused by parameterless benchmarks.
Fixed an issue that caused the benchmark runner to crash and fail to save out results from benchmarks
that'd already completed earlier in the run; the crash was caused by parameterless benchmarks or
comparison functions not producing any Datasets, so we'd get an error when trying to call Dataset.hstack()
on an empty list.

Also added support for capturing versions of selected, relevant packages and saving them
as part of the benchmark metadata.