Fixes #66.
FastArray._reduce_check() failed with a TypeError when given an empty
array and an (explicit) integer dtype. That case is now handled
correctly and we return integer 0 (the identity element, and the same
value numpy returns for this case).
This fixes the case where rt.merge2() with how='outer' failed when
the 'right' Dataset's keyset was a subset of the 'left' Dataset's
keyset. A fast path has also been added for that case so we don't do
the additional work that'd normally be needed (since the outcome would
be the same); effectively, we detect that we can just perform a
LEFT OUTER JOIN instead of a FULL OUTER JOIN.