Parameterize TestCategoricalSimpleGroupBy.test_single_col_categoricals.
Parameterize TestCategoricalSimpleGroupBy.test_single_col_categoricals so
each (function, dtype) pair is run as it's own test case; this makes it
easier to see which pairs are passing and which aren't (instead of halting
on the first failure). Utilize an assertion function from numpy.testing
to provide a richer error message for arrays which aren't equal.