Relax pandas version constraint upper-bound
Users have reported riptable works as expected when interoperating with newer versions of pandas (e.g. 1.1.x, 1.2.x). It seems like we should be able to relax the upper-bound constraint. It'd be good to follow-up by adding a tox configuration to run the unit tests against several of the minor point-release versions of pandas to make sure the riptable <-> pandas interop code works across multiple versions of pandas, e.g. ``0.24``, ``1.0``, ``1.1``, ``1.2``.