Implement unit tests for rt.load_sds() when providing a filter (mask or fancy index).
Implemented additional unit tests for rt.load_sds() when the filter= kwarg is
specified. The existing test which checked the behavior of the function when
a fancy index is supplied for filter= did not check some of these more-complex
scenarios which have come up in real-world use cases. The broken cases are
marked as 'xfail' and will need to be fixed.