Fix FastArray._view_internal() so it works correctly when the output type is not specified.
Implement override of _view_internal() for Categorical.

Implement an override of _view_internal() for Categorical which makes use
of Categorical's .copy() method to make a shallow copy, as opposed to using
the default FastArray._view_internal() implementation which copies over
all object attributes directly (so the original and output instance always
reference the same exact objects, which may not work as intended e.g. when
a Categorical is deleted and it in turn deletes some internal objects).

Fix the ``readonly_array_wrapper()`` helper function so it doesn't break Categoricals.

The ``readonly_array_wrapper()`` helper function is used to create a
read-only view of an array in certain cases in rt.merge2(); specifically,
when an output column doesn't need to be copied so will be referenced
directly in the output Dataset. This was using .view(), which notoriously
breaks Categorical arrays by not copying/cloning their _categories_wrap
and _grouping attributes to the new array object. The function now uses
the FastArray._view_internal() method instead, which knows how to safely
create a view that works for FastArray-derived classes like Categorical.

Implement unit test to demonstrate #136 is fixed.
Utilize the new assert_array_or_cat_equal() function for more-robust/flexible equality checks.

Fixes #136.