Implement override of _view_internal() for Categorical.
Implement an override of _view_internal() for Categorical which makes use
of Categorical's .copy() method to make a shallow copy, as opposed to using
the default FastArray._view_internal() implementation which copies over
all object attributes directly (so the original and output instance always
reference the same exact objects, which may not work as intended e.g. when
a Categorical is deleted and it in turn deletes some internal objects).