Fix the ``readonly_array_wrapper()`` helper function so it doesn't break Categoricals.
The ``readonly_array_wrapper()`` helper function is used to create a
read-only view of an array in certain cases in rt.merge2(); specifically,
when an output column doesn't need to be copied so will be referenced
directly in the output Dataset. This was using .view(), which notoriously
breaks Categorical arrays by not copying/cloning their _categories_wrap
and _grouping attributes to the new array object. The function now uses
the FastArray._view_internal() method instead, which knows how to safely
create a view that works for FastArray-derived classes like Categorical.

Fixes #136.