Disable numba caching due to related segfaults reported by users.
A riptable user ran into a reproducible segfault in the numba-JIT-compiled
code for the fancy_index_fetch_keymap() function in merge, and was able to
definitively trace it back to numba's caching code. The issue is fairly rare
but since it's causing repeatable crashing for some users I'm disabling the
numba caching for riptable functions to provide a short-term (if not
long-term) fix.
This change also introduces a very simple settings/config class that could
be extended with additional configuration settings (to be controlled via
environment variables or some settings file); these would eventually be
used to populate _default_ values for various things like max number of
threads, which we normally auto-detect (and which the user can override),
but this'll allow users to override that auto-detection if they want to.