Fix recently-introduced bug with parallel version of removetrailing.
The recent changes to consolidate numba attributes to make all string functions
have both serial and parallel forms exposed a bug in numba 0.51.2 where the
parallelized version of the `FAString.nb_removetrailing` function caused an
internal compiler error in numba.
Also took the opportunity to re-organize the unit tests alphabetically (by the
name of the function being tested) to make it more obvious which functions have
tests and which don't. Implemented a unit test and test cases for `removetrailing`.