Fix test_ufunc2.py so it's compatible with numpy 1.21.
Instead of using the newest test_ufunc.py from numpy 1.21, the changes were
reverted to the older version; this change makes a small tweak so the older
version of the tests also works with numpy 1.21.

Once we drop support for older versions of numpy, we'll reinstate the
newer version of the file that was reverted.